+++ tiny-lib/TxIn.hpp
@@ -9,14 +9,14 @@
 class TxIn : public ISerializable
 {
 public:
-	TxIn(std::shared_ptr<TxOutPoint> toSpend, const std::vector<uint8_t>& unlockSig, const std::vector<uint8_t>& unlockPk, int64_t sequence);
+	TxIn(std::shared_ptr<TxOutPoint> toSpend, const std::vector<uint8_t>& unlockSig, const std::vector<uint8_t>& unlockPk, int32_t sequence);
 
 	std::shared_ptr<TxOutPoint> ToSpend;
 
 	std::vector<uint8_t> UnlockSig;
 	std::vector<uint8_t> UnlockPk;
 
-	int64_t Sequence;
+	int32_t Sequence;
 
 	std::vector<uint8_t> Serialize() const override;
 };
