+++ tiny-lib/UnspentTxOut.hpp
@@ -10,16 +10,16 @@
 class UnspentTxOut : public ISerializable
 {
 public:
-	UnspentTxOut(int64_t value, const std::string& toAddress, std::shared_ptr<TxOutPoint> txOut, bool isCoinbase, int64_t height);
+	UnspentTxOut(uint64_t value, const std::string& toAddress, std::shared_ptr<TxOutPoint> txOut, bool isCoinbase, int32_t height);
 
-	int64_t Value;
+	uint64_t Value;
 	std::string ToAddress;
 
 	std::shared_ptr<TxOutPoint> TxOut;
 
 	bool IsCoinbase;
 
-	int64_t Height;
+	int32_t Height;
 
 	std::vector<uint8_t> Serialize() const override;
 };
